Feat/gwas unified flow file organization (#1148)
* feat(gwasUnifiedFlowFileOrganization): Created Steps directory and selectStudyPopulation step

* feat(gwasUnifiedFlowFileOrganization): Removed unneeded import from GWASContainer

* feat(gwasUnifiedFlowFileOrganization): Updated code to resolve selection error in SelectStudyPopulation from prior commit

* feat(gwasUnifiedFlowFileOrganization): Fixed ESLINT issues

* feat(gwasUnifiedFlowFileOrganization): Rearranged files into Steps folder

* feat(gwasUnifiedFlowFileOrganization): Renamed CohortSelect to SelectCohort to match 'SelectOutcome'

* feat(gwasUnifiedFlowFileOrganization): Updated Refs to file locations

* feat(gwasUnifiedFlowFileOrganization): Created Components directory

* feat(gwasUnifiedFlowFileOrganization): Removed console statement from reducer.js

* feat(gwasUnifiedFlowFileOrganization): Updated import in ProgressBar.test.jsx and ran npm run eslint-new

* feat(gwasUnifiedFlowFileOrganization): Renamed charts directory to Diagrams

* feat(gwasUnifiedFlowFileOrganization): Fixed imports in storybook file

* feat(gwasUnifiedFlowFileOrganization): Fixed import order in SelectCohort.stories.jsx

* feat(gwasUnifiedFlowFileOrganization): Updated cohortMiddlewareApi location and Searchbar location

* feat(gwasUnifiedFlowFileOrganization): Ran ESLINT-new

* feat(gwasUnifiedFlowFileOrganization): Reverted unintentionally changed file

* feat(gwasUnifiedFlowFileOrganization): Reverted unintentionally changed file (again)

* feat(gwasUnifiedFlowFileOrganization): Reverted unintentionally changed file (third times the charm)
