--- src/Analysis/GWASV2/CohortSelect/Utils/AddCohortButton.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { atlasDomain } from '../../Shared/wizardEndpoints/cohortMiddlewareApi';
-
-const AddCohortButton = () => (
-  <React.Fragment>
-    <button type='button' onClick={() => window.open(atlasDomain(), '_blank')}>
-      Add New Cohort
-    </button>
-  </React.Fragment>
-);
-
-export default AddCohortButton;
