+++ src/Analysis/GWASV2/Components/AttritionTableWrapper/AttritionTable/AttritionTable.jsx
