+++ src/Analysis/GWASV2/Components/AttritionTableWrapper/ChartIcons/BarChart.jsx
