+++ src/Analysis/GWASV2/Components/SelectCohort/Utils/AddCohortButton.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { atlasDomain } from '../../../Shared/cohortMiddlewareApi';
+
+const AddCohortButton = () => (
+  <React.Fragment>
+    <button type='button' onClick={() => window.open(atlasDomain(), '_blank')}>
+      Add New Cohort
+    </button>
+  </React.Fragment>
+);
+
+export default AddCohortButton;
